Main Activity: ActionBar implementation +  ActionBar Listener Creation of GUI AboutActivity, LoginActivity, SettingsActivity (only GUI)
