+++ app/src/main/AndroidManifest.xml
@@ -7,15 +7,18 @@
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme">
+        android:theme="@style/Theme.AppCompat.Light.NoActionBar">
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".TestListActivity"></activity>
+        <activity android:name=".ListSearchActivity" />
+        <activity android:name=".AboutActivity"></activity>
+        <activity android:name=".SettingsActivity"></activity>
+        <activity android:name=".LoginActivity"></activity>
+
     </application>
 
 </manifest>
