--- app/src/main/java/com/jobdirectory/jobdirectory/TestListActivity.java
@@ -1,36 +1,0 @@
-package com.jobdirectory.jobdirectory;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.AdapterView;
-import android.widget.ListAdapter;
-import android.widget.ListView;
-import android.widget.Toast;
-
-public class TestListActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_test_list);
-
-        String[] categories = {"Information technology", "Bank", "Health", "Design", "Hospitality"};
-        ListAdapter buckysAdapter = new CustomAdapter(this, categories);
-        ListView CategoryListView = (ListView) findViewById(R.id.listView1);
-        CategoryListView.setAdapter(buckysAdapter);
-
-
-        CategoryListView.setOnItemClickListener(
-                new AdapterView.OnItemClickListener(){
-                    @Override
-                    public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                        String food = String.valueOf(parent.getItemAtPosition(position));
-                        Toast.makeText(TestListActivity.this, food, Toast.LENGTH_LONG).show();
-                    }
-                }
-        );
-
-    }
-
-}
