+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,26 @@
 <resources>
+
     <string name="app_name">Job directory</string>
+    <string name="action_favorite">Add to my favorites</string>
+    <string name="action_languages">Languages</string>
+    <string name="action_settings">My settings</string>
+    <string name="action_search_reset">Search reset</string>
+    <string name="action_login">Login</string>
+    <string name="action_about">About</string>
+
+    <string name="about_title">A propos</string>
+    <string name="about_copyright">© jobDirectory 2017</string>
+    <string name="about_version">version: 1.1</string>
+    <string name="about_contact">info@jobdirectory.com</string>
+
+    <string name="settings_titre">My settings</string>
+    <string name="settings_languages">Language</string>
+    <string name="settings_languages1">FR</string>
+    <string name="settings_languages2">EN</string>
+    <string name="settings_theme">Theme</string>
+    <string name="settings_theme1">Standard</string>
+    <string name="settings_theme2">Light</string>
+    <string name="settings_btn">Save settings</string>
+
 </resources>
+
